+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
  * Fix missing --url-action and  --browser #113
  * Don't print out URL when sending to browser/clipboard for security
+ * Escape colon in ARN's for `-a` flag to work around the colon being a 
+    word delimiter for bash (auto)complete. #135
 
 ## [v1.3.0] - 2021-11-14